+++ R/tbl-sql.r
@@ -167,6 +167,38 @@ tail.tbl_sql <- function(x, n = 6L, ...) {
   stop("tail is not supported by sql sources", call. = FALSE)
 }
 
+# Set operations ---------------------------------------------------------------
+
+sql_set_op <- function(con, x, y, method) {
+  sql <- build_sql(
+    x$query$sql,
+    "\n", sql(method), "\n",
+    y$query$sql
+  )
+  attr(sql, "vars") <- x$select
+  sql
+}
+
+#' @export
+intersect.tbl_sql <- function(x, y, copy = FALSE, ...) {
+  y <- auto_copy(x, y, copy)
+  sql <- sql_set_op(x$src$con, x, y, "INTERSECT")
+  update(tbl(x$src, sql), group_by = groups(x))
+}
+#' @export
+union.tbl_sql <- function(x, y, copy = FALSE, ...) {
+  y <- auto_copy(x, y, copy)
+  sql <- sql_set_op(x$src$con, x, y, "UNION")
+  update(tbl(x$src, sql), group_by = groups(x))
+}
+#' @export
+setdiff.tbl_sql <- function(x, y, copy = FALSE, ...) {
+  y <- auto_copy(x, y, copy)
+  sql <- sql_set_op(x$src$con, x, y, "EXCEPT")
+  update(tbl(x$src, sql), group_by = groups(x))
+}
+
+
 # SQL select generation --------------------------------------------------------
 
 build_query <- function(x, limit = NULL) {